feat: Implement JWT authentication with Spring Security
- Add JWT dependencies (jjwt 0.12.6)
- Create JwtTokenProvider for token generation/validation
- Create JwtAuthenticationFilter to intercept HTTP requests
- Create UserDetailsServiceImpl for Spring Security
- Create AuthService (login/register) and UserService
- Create AuthController with /api/auth/login and /api/auth/register endpoints
- Update SecurityConfig: enable JWT, stateless sessions, public /api/auth/**
- Add JWT configuration (secret, expiration) to application.properties
- Fix UserRepository ID type: Integer → Long

+++ backend/src/main/java/com/chatop/api/repository/UserRepository.java
@@ -12,19 +12,21 @@ import com.chatop.api.model.User;
  * Provides CRUD operations and custom queries
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    /**
-     * Find a user by email
-     * @param email the user's email
-     * @return Optional containing the user if found
-     */
-    Optional<User> findByEmail(String email);
+  /**
+   * Find a user by email
+   * 
+   * @param email the user's email
+   * @return Optional containing the user if found
+   */
+  Optional<User> findByEmail(String email);
 
-    /**
-     * Check if a user exists by email
-     * @param email the user's email
-     * @return true if user exists, false otherwise
-     */
-    boolean existsByEmail(String email);
+  /**
+   * Check if a user exists by email
+   * 
+   * @param email the user's email
+   * @return true if user exists, false otherwise
+   */
+  boolean existsByEmail(String email);
 }
